requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-aiplatform
pandas
joblib
db-dtypes
scikit-learn
python-dateutil